Conflicts:
	build/cocos2d_libs.xcodeproj/project.pbxproj

+++ cocos/3d/CCMeshSkin.cpp
@@ -70,7 +70,7 @@ ssize_t MeshSkin::getBoneCount() const
 //get bone
 Bone3D* MeshSkin::getBoneByIndex(unsigned int index) const
 {
-    if (index < _skinBones.size())
+    if (static_cast<int>(index) < _skinBones.size())
         return _skinBones.at(index);
     
     return nullptr;
